refactor: removed unnecessary wrapper component
